devcode-agent-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.50.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
ruff==0.6.9
